Fix test coverage of common/tableTools/globalSearchTableTools
